dca-simple/requirements.txt
# Core dependencies for DCA Simple
# Python 3.12 recommended

# OpenAI Agents SDK (includes openai)
git+https://github.com/openai/openai-agents-python.git@73e7843075acaae25bc56628ab8017ae6a4e13f7

# Binance API
python-binance==1.0.32

# Data Analysis (compatible versions)
pandas>=2.0.0,<2.4.0
numpy>=1.23.5,<2.3.0

# Technical Analysis
ta==0.11.0

# Logging & Output
loguru==0.7.3
rich==14.2.0

# HTTP & Async
requests==2.32.5
aiohttp>=3.9.0,<4.0.0

# X (Twitter) API
tweepy==4.14.0

# Environment & Config
python-dotenv==1.2.1
pydantic>=2.0.0,<3.0.0
pydantic-settings>=2.0.0,<3.0.0

# Cryptography (compatible with pyopenssl)
cryptography>=41.0.5,<45.0.0

# Starlette (compatible with fastapi if installed)
starlette>=0.40.0,<0.42.0

# Web Dashboard
Flask>=3.0.0,<4.0.0

# Utilities
python-dateutil==2.9.0.post0
pytz==2025.2
tzlocal==5.3.1
